requirements.txt
colorzero==2.0
paho-mqtt==1.5.1
RPi.GPIO==0.7.0
